IDE inspections

+++ nars_java/src/main/java/nars/Global.java
@@ -101,31 +101,31 @@ public class Global {
 
     /* ---------- default input values ---------- */
     /** Default expectation for confirmation. */
-    public static final float DEFAULT_CONFIRMATION_EXPECTATION = (float) 0.8;
+    public static final float DEFAULT_CONFIRMATION_EXPECTATION = 0.8f;
     /** Default expectation for confirmation. */
-    public static final float DEFAULT_CREATION_EXPECTATION = (float) 0.66;
+    public static final float DEFAULT_CREATION_EXPECTATION = 0.66f;
     /** Default confidence of input judgment. */
-    public static final float DEFAULT_JUDGMENT_CONFIDENCE = (float) 0.9;
+    public static final float DEFAULT_JUDGMENT_CONFIDENCE = 0.9f;
     /** Default priority of input judgment */
-    public static float DEFAULT_JUDGMENT_PRIORITY = (float) 0.8;
+    public static float DEFAULT_JUDGMENT_PRIORITY = 0.8f;
     /** Default durability of input judgment */
-    public static float DEFAULT_JUDGMENT_DURABILITY = (float) 0.5; //was 0.8 in 1.5.5; 0.5 after
+    public static float DEFAULT_JUDGMENT_DURABILITY = 0.5f; //was 0.8 in 1.5.5; 0.5 after
     /** Default priority of input question */
-    public static final float DEFAULT_QUESTION_PRIORITY = (float) 0.9;
+    public static final float DEFAULT_QUESTION_PRIORITY = 0.9f;
     /** Default durability of input question */
-    public static final float DEFAULT_QUESTION_DURABILITY = (float) 0.9;
+    public static final float DEFAULT_QUESTION_DURABILITY = 0.9f;
 
     
      /** Default confidence of input goal. */
-     public static final float DEFAULT_GOAL_CONFIDENCE = (float) 0.9;
+     public static final float DEFAULT_GOAL_CONFIDENCE = 0.9f;
      /** Default priority of input judgment */
-     public static final float DEFAULT_GOAL_PRIORITY = (float) 0.9;
+     public static final float DEFAULT_GOAL_PRIORITY = 0.9f;
      /** Default durability of input judgment */
-     public static final float DEFAULT_GOAL_DURABILITY = (float) 0.9;
+     public static final float DEFAULT_GOAL_DURABILITY = 0.9f;
      /** Default priority of input question */
-     public static final float DEFAULT_QUEST_PRIORITY = (float) 0.9;
+     public static final float DEFAULT_QUEST_PRIORITY = 0.9f;
      /** Default durability of input question */
-     public static final float DEFAULT_QUEST_DURABILITY = (float) 0.9;
+     public static final float DEFAULT_QUEST_DURABILITY = 0.9f;
  
     
     /* ---------- space management ---------- */
